requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
pymongo==4.8.0
python-dotenv==1.0.1
google-generativeai==0.7.2
python-docx==1.1.2
python-pptx==0.6.23
bcrypt==4.1.3
Werkzeug==3.0.3